pyproject.toml
[tool.poetry]
name = "hard-brain-bot"
version = "0.1.0"
description = ""
authors = ["Nathan <101812777+corndogit@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.1"
disnake = "^2.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
